requirements.txt
requests==2.31.0
PyYAML==6.0.1
schedule==1.2.0
pymongo==4.5.0
plotly==5.18.0